% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.report_FIBION.R
\name{get.report_FIBION}
\alias{get.report_FIBION}
\title{get.report_FIBION}
\usage{
get.report_FIBION(
  datadir = NULL,
  data = NULL,
  outputdir = "./",
  store.csv = FALSE
)
}
\arguments{
\item{datadir}{Path to directory where the output csv files exported from FIBION are stored.}

\item{outputdir}{Path to directory where the data sets should be stored.}

\item{store.csv}{Logical to indicate whether the csv reports should be stored as csv files in outputdir.}
}
\value{
FIBION data aggregated at day and week levels.
}
\description{
get.report_FIBION
}
