% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.report_FIBION.R
\name{get.report_FIBION}
\alias{get.report_FIBION}
\title{get.report_FIBION}
\usage{
get.report_FIBION(
  datadir = NULL,
  data = NULL,
  outputdir = "./",
  store.csv = FALSE,
  ID = NULL
)
}
\arguments{
\item{datadir}{Character. Path to directory where the output csv files exported from FIBION are stored.}

\item{outputdir}{Character. Path to directory where the data sets should be stored.}

\item{store.output}{Logical (default = FALSE). Indicate whether the generated datasets should be stored in outputdir.}

\item{ID}{Numeric or character (default = NULL). Specify the ID of the file being processed (only works if processing 1 file).}
}
\value{
FIBION data aggregated at day and week levels.
}
\description{
get.report_FIBION
}
